% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_CSDR_1921_1.R
\name{add_cols_CSDR_1921_1}
\alias{add_cols_CSDR_1921_1}
\title{add_cols_CSDR_1921_1}
\usage{
add_cols_CSDR_1921_1(CSDR_1921_1)
}
\arguments{
\item{CSDR_1921_1}{Required. A tibble of CSDR 1921-1 data that has been
imported using the \code{import_CSDR_1921()} function.}
}
\value{
Returns a modified tibble of the CSDR 1921-1 data with additional
columns added.
}
\description{
Add Supporting Columns to a CSDR 1921-1 Tibble
}
\details{
Takes a CSDR 1921-1 tibble and adds additional columns to allow for
additional/easier sorting/filtering.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
