% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_CSDR_Detailed_Definitions.R
\docType{data}
\name{detailed_defs}
\alias{detailed_defs}
\title{CSDR Detailed Definitions}
\format{A tibble of factors with 93 rows and 8 columns:
\describe{
\item{\code{Short Name}}
\describe{
\item{test1}{test1}
\item{test2}{test2}}
\item{\code{Full Name}}{}
\item{\code{Unit of Measure}}{}
\item{\code{To Date / At Completion}}{}
\item{\code{Recurring / Nonrecurring}}{}
\item{\code{Functional Category}}{}
\item{\code{Functional Element}}{
\describe{
\item{\code{Direct Engineering Labor (Hours and Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to Acquisition efforts only. Engineering includes the scientific exploration, study, analysis, design, development, evaluation, and redesign of a specific task or WBS element. Engineering also includes preparation of specifications, drawings, parts lists, and wiring diagrams; technical coordination between engineering and manufacturing; design of tools; coordination of suppliers; planning for and scheduling of tests; analysis of test results; reduction of data; and preparation of reports. It also includes the determination and specification of requirements for reliability, maintainability, and quality control, as well as logistics and training engineering efforts such as training system and manual development, support equipment design and development, logistics support analysis and other support engineering efforts.}}
\item{\code{Engineering Labor Overhead (Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to both Acquisition and Sustainment efforts. It includes all indirect engineering costs and fringe, but excludes G&A expenses. For example, Engineering Overhead includes the cost of directing and supporting all Engineering organization-related activities that cannot be assigned to specific contracts. Engineering overhead typically includes supervision, policy and procedures, training, administration, and similar costs.}}
\item{\code{Direct Manufacturing Other Labor (Hours and Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to Acquisition efforts only. The effort expended to manufacture a product or end item that meets any of the Tier 2 definitions below.}
\describe{
\item{\code{Direct Manufacturing Support Labor (Hours and Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to Acquisition efforts only. It includes support tothe manufacturing and maintenance processes such as planning, material handling, qualitycontrol, and tool maintenance. Quality control entails checking, physically inspecting,measuring, testing, or otherwise verifying that products and services conform toestablished technical requirements and that satisfactory performance is achieved}}
\item{\code{Direct Manufacturing Tooling Labor (Hours and Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to Acquisition efforts only. MManufacturing Tooling is a labor cost element of the Manufacturing Operations functional category that includes the effort and costs expended to acquire, manufacture, maintain, or replace original equipment and manufacturing aids. It does not include the cost of procuring tooling materials or equipment from an external supplier.}}
\item{\code{Direct Manufacturing Other Labor (Hours and Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to Acquisition efforts only. Entries into this category may pertain to Direct Manufacturing Touch, Support, or Tooling Labor if the reporting entity cannot clearly segregate those elements. It may also contain Direct Manufacturing labor elements that do not match the Touch, Support, or Tooling elements above.}}
}
}
\item{\code{Direct Manufacturing Touch Labor (Hours and Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to Acquisition efforts only. It includes the fabrication, assembly, integration, application of paint and coatings, and functional testing of a product or end item, andmay include certain types of quality control.}}
\item{\code{Manufacturing Operations Labor Overhead (Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to Acquisition efforts only. This category covers Direct Manufacturing Touch and Support Labor functions. Manufacturing Operations Overhead consists of all indirect costs, including fringe, but excluding G&A expenses. For example, Manufacturing Operations Overhead includes the cost of directing and supporting all Manufacturing-organizationrelated activities that cannot be assigned to specific contracts. Manufacturing Operations Overhead typically includes supervision, policy and procedures, training, administration, and similar costs}}
\item{\code{Direct Maintenance Other Labor (Hours and Dollars)}}{\code{This is a Tier 1 (required) StandardFunctional Category applicable to Sustainment efforts only. Direct Maintenance Operations Labor is the effort expended to maintain a product or end item that meets any of the Tier 2 definitions below.}
\describe{
\item{\code{Direct Maintenance Support Labor (Hours and Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to Sustainment efforts only. Maintenance support functions include planning, material handling, tool maintenance, and quality control efforts.}}
\item{\code{Direct Maintenance Other Labor (Hours and Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to Sustainment efforts only. Entries into this category may pertain to Direct Maintenance Touch or Support Labor if the reporting entity cannot clearly segregate those elements. It may also contain Direct Maintenance labor elements that do not match the Touch or Support elements above.}}
}
}
\item{\code{Direct Maintenance Touch Labor (Hours and Dollars)}}{\code{This is a Tier 1 (Required) Standard Functional Category applicable to Sustainment efforts only. It includes the effort and costs expended for hands-on post-production maintenance of final prime mission product items, and may include certain types of quality control.}}
\item{\code{Maintenance Operations Labor Overhead (Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to Sustainment efforts only. It covers Direct Maintenance Touch and Maintenance Support Labor functions. Maintenance Operations Overhead consists of all indirect costs, including fringe, but excluding G&A expenses.}}
\item{\code{Other Direct Costs Not Shown Elsewhere (Hours and Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to both Acquisition and Sustainment efforts. Direct costs not assigned to the Engineering, Manufacturing Operations, Maintenance, or Materials functions are included in Other Direct Costs Not Shown Elsewhere. This category includes labor and non-labor costs that meet any of the Tier 2 definitions below.}
\describe{
\item{\code{Direct Program Management Labor (Hours and Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. It includes the business and administrative planning, organizing, directing, coordinating, controlling, and approval actions designated to accomplish overall program objectives, which are not associated with specific hardware elements and are not included in systems engineering.}}
\item{\code{Direct Other Labor (Hours and Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. It includes any labor category that does not align with the engineering, manufacturing operations, maintenance, or program management functional cost categories. It may include Direct Program Management labor if the reporting entity cannot clearly segregate those elements, as well as any other direct labor costs incurred by the reporting entity that does not match the definitions listed above.}}
\item{\code{Direct Services (Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Categoryapplicable to both Acquisition and Sustainment efforts. It includes subcontracted services not delivering a physical product that are a direct charge to the program. Examples might include information technology support, engineering support, consultants, etc. Do not include services that are an overhead cost.}}
\item{\code{Other Direct Non-Labor (Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. It includes direct non-labor costs such as travel, per diem, reproduction of printed material, rental of special facilities and equipment, and shipping and transportation charges for items sent or returned to subcontractors. It may include Direct Services if the reporting entity cannot clearly segregate those elements, as well as any other direct non-labor costs incurred by the reporting entity that does not match the definitions listed above or below.}}
}
}
\item{\code{Other Overhead (Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to both Acquisition and Sustainment efforts. Other Overhead includes all functions tagged to Other Direct Costs Not Shown Elsewhere, including Program Management, Services, Travel, etc. Other Overhead consists of all indirect costs, including fringe, but excluding G&A expenses.}}
\item{\code{Direct Materials (Dollars)}}{\code{This is a Tier 1 (required) Standard Functional Category applicable to both Acquisition and Sustainment efforts. Direct Materials is defined as the dollars paid for material items that meet any of the Tier 2 definitions below.}
\describe{
\item{\code{Direct-Reporting Subcontractor (Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. It includes the total price of all direct-reporting subcontracts. Subcontractor costs that do not directly report a FlexFile should be reported in one of the other material categories below based upon the nature and characteristics of the product provided.}}
\item{\code{Intercompany Work Orders (IWO) (Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. An IWO is work completed by a separate business unit of the reporting entity.}}
\item{\code{Purchased Parts (Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. Purchased Parts are cost elements of the Materials functional category that includes items that are discrete components used in an upper-level assembly. Purchased Parts are distinguished from purchased equipment by their relatively lower cost and complexity. Examples include fasteners, clips, clamps, nuts, bolts, washers, nails, screws, valves, and plumbing and electrical fittings and fixtures.}}
\item{\code{Purchased Equipment (Dollars)}}{\code{ This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. Purchased Equipment is a cost element of the Materials functional category that includes assembled items designed to be incorporated with other components into a finished product. Purchased Equipment is distinguished from Purchased Parts by its relatively higher cost and complexity. Aviation examples include structural components such as wings, horizontal and vertical tails, and fuselage; avionics equipment such as radios, inertial navigation systems, radar systems, and electronic countermeasures; and hydraulic, pneumatic, and electrical subassemblies such as landing gear, canopy actuation systems, and wire harnesses. Other examples include brake disks, tires, transmission, propellers, lenses, scopes, and dishes.}}
\item{\code{Raw Materials (Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. Raw Materials are cost elements of the Materials functional category that includes items that are crude, semi-fabricated, or partially processed materials or components that have not yet been made into a definite functional item or configuration. Examples include consumable items for fabrication, castings, forgings, pressings, sheet metal, plate, tubing, bars, rebar, rods, wires, cables, fabrics, and conduits.}}
\item{\code{Direct Tooling and Equipment (Dollars)}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. It includes the cost associated with labor, materials, and equipment used in the manufacture of dies, jigs, fixtures, molds, gauges, handling equipment, work platforms, and test equipment for the fabrication and testing of the specific WBS reporting element.}}
\item{\code{Other Material (Dollars):}}{\code{This is a Tier 2 (optional) Standard Functional Category applicable to both Acquisition and Sustainment efforts. Other Materials is a cost element of the Materials functional category that may pertain to Direct-Reporting Subcontractor, IWO, Purchased Parts, Purchased Equipment, Raw Materials, or Tooling and Equipment if the reporting entity cannot clearly segregate those elements. It may also contain other Direct Material elements that do not match any of the Direct Materials definitions for Tier 2 above.}}
}
}
\item{\code{Material Overhead (Dollars)}}{\code{ This is a Tier 1 (required) Standard Functional Category applicable to both Acquisition and Sustainment efforts. All indirect material costs relating to Direct Materials (including Direct-Reporting Subcontractors and IWO), but excluding G&A expenses. For example, the portion of indirect costs attributable to procured or subcontracted products, including the cost of purchasing, expediting, and storing materials, parts, equipment, and assemblies.}}
\item{\code{General and Administrative (G&A) (Dollars)}}{This is a Tier 1
(required) Standard Functional Category applicable to both Acquisition and
Sustainment efforts. G&A costs are indirect expenses related to the overall
management and administration of the contractor’s business unit, including
the following: a company’s general and executive offices; the cost of staff
services such as legal, accounting, public relations, financial, and
similar expenses; and other general expenses. G&A is also a generic term
used to describe expenses with a beneficial or causal relationship to cost
objectives that cannot be more accurately assigned to overhead areas for
Engineering, Manufacturing Operations, Material, and similar costs.

\strong{\emph{The reporting entity has the choice of whether to report G&A
costs along with the Actuals To Date information, or to withhold G&A
entirely from this Data Group; if the reporting entity elects to report
G&A, then it must be applied for all direct-charges and identification of
G&A as a functional category is required.}}}
\item{\code{Facilities Capital Cost of Money (FCCOM) (Dollars)}}{This is a Tier 1
(required) Standard Functional Category applicable to both Acquisition and
Sustainment efforts. FCCOM is an imputed cost determined by applying a ‘cost
of money’ rate to facilities capital employed in contract performance
according to Cost Accounting Standard 414, “Cost of Money as an Element of
the Cost of Facilities Capital.” Capital employed is determined without
regard to whether its source is equity or borrowed capital. The resulting
cost of money is not a form of interest on borrowing.

\strong{\emph{The reporting entity has the choice of whether to report FCCOM costs along
with the Actuals To Date information, or to withhold FCCOM entirely from this
Data Group; if the reporting entity elects to report FCCOM, then it must be
applied for all direct-charges and identification of FCCOM as a functional
category is required.}}}}}
\item{\code{Form}}{The CSDR DD Form type the data field is used in.}
}}
\usage{
detailed_defs
}
\description{
The CSDR "Detailed Definitions" csv file which is exported when doing a
cross platform query, in CADE, converted into a tibble for easier use in R.
}
\keyword{datasets}
