% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_import_functions.R
\name{.create_excel_ws_ls}
\alias{.create_excel_ws_ls}
\title{.create_excel_ws_ls}
\usage{
.create_excel_ws_ls(folder_path, recurse = TRUE)
}
\arguments{
\item{folder_path}{Required. A path to the folder with Excel files you want
data on.}

\item{recurse}{Optional. If \code{TRUE} recurse fully, if a positive number
the number of levels to recurse. Set to \code{TRUE} by default.}
}
\value{
Returns a tibble with the following columns of data on all ".xls\link{x}"
files found at the provided path: \enumerate{ \item path: A full path to
each Excel file found at the provided folder path. \item folder_path: The
folder path the Excel file was found in. \item file_name: The Excel file
name (with extention) found in the folder. \item worksheet_title: The title
of each worksheet in the Excel file.}
}
\description{
Create an Excel File and Worksheet Table
}
\details{
Creates a tibble of information on all Excel (".xls\link{x}") files at the
provided path.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
