% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_to_excel.R
\name{.send_to_excel}
\alias{.send_to_excel}
\title{Send to Excel}
\usage{
.send_to_excel(x)
}
\arguments{
\item{x}{Required. The (presumably) table object to send to Excel.}
}
\value{
Opens an Excel file with the object on the first worksheet.
}
\description{
A simple wrapper function used to send a single object to a temp excel file.
}
