% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstCSDR.R
\docType{package}
\name{dstCSDR}
\alias{dstCSDR}
\alias{dstCSDR-package}
\title{\code{dstCSDR} package}
\description{
Tools for Cost and Software Data Reporting (CSDR) Data
}
\details{
See the README on
or \href{https://github.com/daniel-r-germony/dstCSDR#readme}{GitHub}
}
