% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstCSDR_files.R
\name{dstCSDR_files}
\alias{dstCSDR_files}
\title{Get the path to, or a list of, \code{{dstCSDR}} file(s)}
\usage{
dstCSDR_files(path = NULL)
}
\arguments{
\item{path}{Name of a specific \code{{dstCSDR}} file or \code{NULL}.}
}
\value{
If \code{path} is \code{NULL} (the default), a list of all files will be
returned. If \code{path} is a specific file name (with extension) is
provided, the full file path is returned.
}
\description{
\code{{dstCSDR}} comes with various CSDR related files in the package's \code{extdata}
directory. This function is used to return a list of the files or to specify
one for you may want to use.
}
\examples{

# Lists all the {dstCSDR} files.
dstCSDR_files()

# Return a full path to the "Example_CSDR_1921.xls" file.
dstCSDR_files("Example_CDSR_1921.xls")

\dontrun{

# Copy "cost_reporting_params.xlsx" and put it at the `to = "..."` location.
dstCSDR_files("cost_reporting_params.xlsx") \%>\%
  file.copy(to = "C:/Users/john.doe/Desktop")
}
}
