% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dstCSDR_files.R
\name{example_dstCSDR_files}
\alias{example_dstCSDR_files}
\title{Get the path to, or a list of, dstCSDR example file(s)}
\usage{
example_dstCSDR_files(path = NULL)
}
\arguments{
\item{path}{Name of a specific dstCSDR example file or \code{NULL}.}
}
\value{
If \code{path} is \code{NULL} (the default), a list of all example files
will be returned. If \code{path} is a specific file name (with extension)
is provided, the full file path is returned.
}
\description{
dstCSDR comes with various CSDR related example files in the package's
\code{extdata} directory. This function is used to return a list of the example
files or to specify one for you may want to use as an example.
}
\examples{
example_dstCSDR_files()
example_dstCSDR_files("Example_CDSR_1921.xls")
}
