% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_fchr.R
\name{gather_fchr}
\alias{gather_fchr}
\title{gather_fchr}
\usage{
gather_fchr(fchr_object)
}
\arguments{
\item{fchr_object}{Required. A list object of CSDR FCHR 1921-1 data that has
been imported using the \code{import_cdsr_excel()} function.}
}
\value{
Returns a CSDR FCHR (DD Form 1921-1) object with all "Cost and
Hours..." columns converted from columns to rows in a new "Reported Data
Field" column.
}
\description{
Gather CDSR FCHR (DD Form 1921-1)
}
\details{
Gathers the "Cost and Hours..." columns from a CSDR Functional Cost-Hour
Report (FCHR) (DD Form 1921-1) object in a new column called "Reported Data
Field". This longer and narrower version of the FCHR is eaiser to
pivot/sort/filter.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
