% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_CSDR_1921_XML.R
\name{get_files_ccdr}
\alias{get_files_ccdr}
\title{Get CCDR Files from a Directory}
\usage{
get_files_ccdr(file_path)
}
\arguments{
\item{file_path}{The root directory to search from.}
}
\value{
A 6 column table (tbl_df) of XML files and their type (e.g., 1921, 1921-1).
}
\description{
For a given directory (i.e., file path), return a recursive list of all CSDR XML files.
}
\examples{
\dontrun{\donttest{
file_path <- system.file("extdata", package = "dstCSDR")
get_files_ccdr(file_path)
}}
}
