% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cdsr_excel.R
\name{import_cdsr_excel}
\alias{import_cdsr_excel}
\title{import_cdsr_excel}
\usage{
import_cdsr_excel(path)
}
\arguments{
\item{path}{Required. Path to the xls/xlsx file which contains the CSDR
1921. Note the Excel file must have the CSDR 1921 data on the first
worksheet (remove any coversheets prior to using this script) and cannot
have columns or rows moved/changed from the standard from (i.e., the
script assumes a cPet produced/compliant file is provided).}
}
\value{
Returns a list of three tibbles. The first tibble includes all of
the submission's metadata in two columns. The second tibble includes each
WBS element as a row and the repoted data as columns. The third tibble
includes the summary reporting elements (the rows at the bottom of a
CDSR with summary cost data and no WBS number).
}
\description{
Import a CSDR CDSR (DD Form 1921) from Excel into a Tibble
}
\details{
Takes a CSDR CDSR (DD Form 1921) Excel worksheet and imports it to three a
tibbles.
}
\examples{
# Import the example CDSR.
import_cdsr_excel(dstCSDR_files("Example_CDSR_1921.xls"))

\dontrun{

# Import a CDSR from a fake user desktop by directly passing the path.
import_cdsr_excel("C:/Users/john.doe/Desktop/DD Form 1921 submission.xls")

# Import a CDSR from a fake user desktop by passing a path object.
cdsr_path <- "C:/Users/john.doe/Desktop/DD Form 1921 submission.xls"
import_cdsr_excel(cdsr_path)

}
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
