% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSDR_basic_import.R
\name{import_CSDR}
\alias{import_CSDR}
\title{import_CSDR}
\usage{
import_CSDR(path_CSDR_1921, path_CSDR_1921_1)
}
\arguments{
\item{path_CSDR_1921}{Required. Path to the xls/xlsx file which contains
the CSDR 1921. Note the Excel file must have the CSDR 1921 data on the
first worksheet (remove any coversheets prior to using this script) and
cannot have columns or rows moved/changed from the standard from (i.e.,
the script assumes a cPet produced/compliant file is provided).}

\item{path_CSDR_1921_1}{Required. Path to the xls/xlsx file which contains
the CSDR 1921-1 data. Note the Excel file must include only CSDR 1921-1
worksheets, one worksheet per WBS (remove any coversheets prior to using
this script) and cannot have columns or rows moved/changed from the
standard from (i.e., the script assumes a cPet produced/compliant file
is provided).}
}
\value{
Returns a list with three objects, one of the CSDR 1921 tibble, one
with the CSDR 1921-1 tibble, and one with the CSDR 1921-1+ tibble (i.e.,
the CSDR 1921-1 tibble with additional columns.
}
\description{
Import a CSDR 1921 and 1921-1
}
\details{
Takes two paths to Excel files (one for a CSDR 1921 and one for a 1921-1)
and runs the three primary \code{dstCSDR} package import functions:
\enumerate{
\item \code{import_CSDR_1921()}
\item \code{import_CSDR_1921-1()}
\item \code{add_cols_CSDR_1921_1()}
}
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
