% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_CSDR_1921.R
\name{import_CSDR_1921}
\alias{import_CSDR_1921}
\title{import_CSDR_1921}
\usage{
import_CSDR_1921(path)
}
\arguments{
\item{path}{Required. Path to the xls/xlsx file which contains the CSDR
1921. Note the Excel file must have the CSDR 1921 data on the first
worksheet (remove any coversheets prior to using this script) and cannot
have columns or rows moved/changed from the standard from (i.e., the
script assumes a cPet produced/compliant file is provided).}
}
\value{
Returns a tibble with each WBS element as a row and traditional
CSDR 1921 data as columns.
}
\description{
Import a CSDR 1921 from Excel into a Tibble
}
\details{
Takes a CSDR 1921 Excel worksheet and imports it to a tibble.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
