% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_CSDR_1921_1.R
\name{import_CSDR_1921_1}
\alias{import_CSDR_1921_1}
\title{import_CSDR_1921_1}
\usage{
import_CSDR_1921_1(path)
}
\arguments{
\item{path}{Required. Path to the xls/xlsx file which contains the CSDR
1921-1 data. Note the Excel file must include  only CSDR 1921-1
worksheets, one worksheet per WBS (remove any coversheets prior to using
this script) and cannot have columns or rows moved/changed from the
standard from (i.e., the script assumes a cPet produced/compliant file
is provided).}
}
\value{
Returns a tibble with each CSDR 1921-1 worksheet combined into one
table and traditional  CSDR 1921-1 data as columns. A limited amount of
transformations are done to make produce a proper table of data (e.g.,
functional categories are converted from padded rows to a column).
}
\description{
Import a CSDR 1921-1 from Excel into a Tibble
}
\details{
Takes a CSDR 1921-1 Excel file and imports it to a single long tibble.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
