% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ff_flatfile.R
\name{import_flexfile_flatfile}
\alias{import_flexfile_flatfile}
\title{import_ff_flatfile}
\usage{
import_flexfile_flatfile(path)
}
\arguments{
\item{path}{Required. Path to the csv file which contains the CSDR FlexFile
flat file.}
}
\value{
Returns a tibble of the CSDR FlexFile data.
}
\description{
Import a CSDR FlexFile Flat File
}
\details{
Takes a CSDR FlexFile Flat File csv and imports it to a tibble.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
