% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wider_CSDR_1921_1.R
\name{pivot_wider_CSDR_1921_1}
\alias{pivot_wider_CSDR_1921_1}
\title{pivot_wider_CSDR_1921_1}
\usage{
pivot_wider_CSDR_1921_1(CSDR_1921_1_plus, ...)
}
\arguments{
\item{CSDR_1921_1_plus}{Required. A tibble of CSDR 1921-1 data that has
been imported using the \code{import_CSDR_1921()} function and run thru the
\code{add_cols_CSDR_1921_1()} function.}

\item{...}{Optional. Allows for use of \code{filter()} from \code{{dplyr}} and other piped functions.}
}
\value{
Returns a tibble with each WBS element as a row and CSDR 1921-1
data as columns.
}
\description{
Transpose CSDR 1921-1 Data
}
\details{
Takes CSDR 1921-1 data and performs a transpose/pivot_wider so each row is
a WBS and each column is one of the CSDR 1921-1's data cells. The primary
use case for this is to lay CSDR 1921-1 data alongside a CSDR 1921 submission.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
