% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_CSDR_1921_XML.R
\name{read_xml_ccdr}
\alias{read_xml_ccdr}
\title{Read a DD Form 1921 XML File}
\usage{
read_xml_ccdr(file_path)
}
\arguments{
\item{file_path}{The file to read.}
}
\value{
A list of three tables (tbl_df): \cr
\describe{
\item{metadata}{The document metadata (e.g., Program Name, Contract Number, Report As Of Date, etc.).}
\item{cost_report}{The WBS level cost report data.}
\item{cost_summary}{The summary level cost report data (e.g., Subtotal, G&A, Fee).}
}
}
\description{
Parse the data for a given file name with a valid extension (i.e., '1921.xml').
}
\examples{
\dontrun{\donttest{
file_path <- system.file("extdata", package = "dstCSDR")
get_files_ccdr(file_path)

cost_report_1 <- read_xml_ccdr(paste0(file_path, ccdr_files$file[1]))
}}
}
