% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs_code_to_lvl.R
\name{wbs_code_to_lvl}
\alias{wbs_code_to_lvl}
\title{WBS Element Code to Level}
\usage{
wbs_code_to_lvl(wbs_element_code, ...)
}
\arguments{
\item{wbs_element_code}{Required.}

\item{...}{Optional. Included in order to be %>% frendly.}
}
\value{
Returns an intiger assoicated with WBS Element Code (i.e., 0 for WBS
= "1.0", 1 for "1.1", 3 for "1.2.2.1")
}
\description{
A function to identify the level assoicated with given WBS Element Code.
}
