% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen_fchr.R
\name{widen_fchr}
\alias{widen_fchr}
\title{widen_fchr}
\usage{
widen_fchr(fchr_plus_object, ...)
}
\arguments{
\item{fchr_plus_object}{Required. A list object of CSDR FCHR 1921-1 data that
has been imported using \code{import_cdsr_excel()} and processed thru
\code{add_supplemental_fchr_columns()}.}

\item{...}{Optional. Allows for use of \code{filter()} from \code{{dplyr}}
and other piped functions.}
}
\value{
Returns a tibble with each WBS element as a row and CSDR 1921-1 data
as columns.
}
\description{
Transpose CSDR FCHR (DD Form 1921-1) Data
}
\details{
Takes CSDR FCHR (DD Form 1921-1) data and performs a transpose/pivot_wider so
each row is a WBS and each column is one of the CSDR FCHR's (DD Form
1921-1's) data cells. The primary use case for this is to lay CSDR FCHR (DD
Form 1921-1) data alongside a CSDR CDSR (DD Form 1921) submission.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
