% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_supplemental_fchr_columns.R
\name{add_supplemental_fchr_columns}
\alias{add_supplemental_fchr_columns}
\title{add_supplemental_fchr_columns}
\usage{
add_supplemental_fchr_columns(fchr_object)
}
\arguments{
\item{fchr_object}{Required. A list object of CSDR FCHR 1921-1 data that has
been imported using the \code{import_cdsr_excel()} function.}
}
\value{
Adds the following columns to the \code{`reported_data`} tibble:
\enumerate{
\item Functional Category
\item Recurring / Nonrecurring
\item To Date / At Completion
\item Functional Element
\item Functional Data Element Number
}
}
\description{
Add Supplemental Columns to a CDSR FCHR (DD Form 1921-1) Tibble
}
\details{
Takes a CSDR Functional Cost-Hour Report (FCHR) (DD Form 1921-1) object and
adds supplemental columns to the to the \code{`reported_data`} tibble to
allow for additional/easier sorting/filtering.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
