% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_supplemental_fchr_columns.R
\name{add_supplemental_fchr_columns}
\alias{add_supplemental_fchr_columns}
\title{add_supplemental_fchr_columns}
\usage{
add_supplemental_fchr_columns(fchr_object)
}
\arguments{
\item{fchr_object}{Required. A tibble of CSDR 1921-1 data that has been
imported using the \code{import_CSDR_1921()} function.}
}
\value{
Returns a modified tibble of the CSDR 1921-1 data with additional
columns added.
}
\description{
Add Supplemental Columns to a CDSR FCHR (DD Form 1921-1) Tibble
}
\details{
Takes a CSDR Functional Cost-Hour Report (FCHR) (DD Form 1921-1) tibble and
adds supplemental columns to allow for additional/easier sorting/filtering.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
