% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excel_ws_ls.R
\name{.create_excel_ws_ls}
\alias{.create_excel_ws_ls}
\title{.create_excel_ws_ls}
\usage{
.create_excel_ws_ls(folder_path, recurse = TRUE)
}
\arguments{
\item{folder_path}{Required. A path to the folder with Excel files you want
data on.}

\item{recurse}{Optional. If \code{TRUE} recurse fully, if a positive number
the number of levels to recurse. Set to \code{TRUE} by default.}
}
\value{
Returns a tibble with the following columns of data on all ".xls" or
".xlsx" files found at the provided path: \enumerate{ \item \code{path}: A
full path to each Excel file found at the provided folder path. \item
\code{folder_path}: The folder path the Excel file was found in. \item
\code{file_name}: The Excel file name (with extention) found in the folder.
\item \code{worksheet_title}: The title of each worksheet in the Excel
file.}
}
\description{
Create an Excel File and Worksheet Tibble
}
\details{
Creates a tibble of information on all Excel (".xls" or ".xlsx") files found
in a provided folder path.
}
\seealso{
\code{\link[=.import_excel_ws_ls]{.import_excel_ws_ls()}} for how to import worksheets identfied in
the tibble created with this function.

\code{\link[=.grab_cell]{.grab_cell()}} pulls the value out of an individual Excel cell and
reurns it to R.

Other dstCSDR Import Excel Data Functions: 
\code{\link{.grab_cell}()},
\code{\link{.import_excel_ws_ls}()}
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
\concept{dstCSDR Import Excel Data Functions}
