% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_cell.R
\name{.grab_cell}
\alias{.grab_cell}
\title{Get Excel Cell Value}
\usage{
.grab_cell(path, cell_range)
}
\arguments{
\item{path}{Required. The full path to an Excel file.}

\item{cell_range}{Required. The cell range where you want to pull data form.
Must be in the "A1" style.}
}
\value{
Returns the value from the Excel file as an "NA" if the cell was
blank or as a character otherwise
}
\description{
A function which pulls the value out of an individual Excel cell and reurns
it to R.
}
\seealso{
\code{\link[=.create_excel_ws_ls]{.create_excel_ws_ls()}} to create a tibble of files and worksheet
names.

\code{\link[=.import_excel_ws_ls]{.import_excel_ws_ls()}} for how to import worksheets identfied in
the tibble created with this function.

Other dstCSDR Import Excel Data Functions: \code{\link{.create_excel_ws_ls}},
  \code{\link{.import_excel_ws_ls}}
}
\concept{dstCSDR Import Excel Data Functions}
