% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel_ws_ls.R
\name{.import_excel_ws_ls}
\alias{.import_excel_ws_ls}
\title{.import_excel_ws_ls}
\usage{
.import_excel_ws_ls(excel_ws_ls)
}
\arguments{
\item{excel_ws_ls}{Required. A tibble object created (preferablly) using
\code{\link[dstCSDR:dot-create_excel_ws_ls]{.create_excel_ws_ls}} which has
a column of file paths and worksheet titles that will be imported.}
}
\value{
Returns a tibble with each worksheet listed in the \code{excel_ws_ls}
object combined into one tibble. A column with \code{source_file_name} and
\code{source_file_worksheet} are also added to keep track of where the data
came from.
}
\description{
Import Worksheets From a \code{excel_ws_ls} Object
}
\details{
Imports all worksheets listed in a \code{excel_ws_ls} object to a single
tibble. Each column in the source worksheets should be of identical names and
types.
}
\seealso{
\code{\link[=.create_excel_ws_ls]{.create_excel_ws_ls()}} to create a tibble of files and worksheet
names.

\code{\link[=.grab_cell]{.grab_cell()}} pulls the value out of an individual Excel cell and
reurns it to R.

Other dstCSDR Import Excel Data Functions: \code{\link{.create_excel_ws_ls}},
  \code{\link{.grab_cell}}
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
\concept{dstCSDR Import Excel Data Functions}
