% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_import_functions.R
\name{.import_excel_ws_ls}
\alias{.import_excel_ws_ls}
\title{.import_excel_ws_ls}
\usage{
.import_excel_ws_ls(excel_ws_ls)
}
\arguments{
\item{excel_ws_ls}{Required. A tibble object created (preferablly) using
\code{.create_excel_ws_ls} which has a column of file paths and worksheet
titles that will be imported.}
}
\value{
Returns a tibble with each worksheet listed in the \code{excel_ws_ls}
object combined into one tibble. A column for \code{source_file_name} and
\code{source_file_worksheet} are also added to keep track of where the data
came from.
}
\description{
Import Worksheets From Given excel_ws_ls Object
}
\details{
Imports all worksheets listed in a excel_ws_ls object to a single tibble.
Each column in the source worksheets should be of identical names and types.
}
\author{
Daniel Germony \email{daniel.r.germony.civ@mail.mil}
}
