% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_cr_docs.R
\name{render_cr_docs}
\alias{render_cr_docs}
\title{Create Cost Reporting SOW, CDRLs, Attachments with Params}
\usage{
render_cr_docs(params_path = file.choose(), output_dir = choose.dir(),
  files_to_render = "default")
}
\arguments{
\item{params_path}{Required. A full path to a filled-in
"cost_reporting_params.xlsx" file. If a path is not provided to the
function call, a file selection box will be opened.}

\item{output_dir}{Required. A directory path to where the user wants the
generated files to be placed. If using Windows, a pop-up box will ppear to
specify an output directory. Non-Windows users must pass a directory path
in the function call.}

\item{files_to_render}{Optional. Specify which files are rendered and how
they are returned. Options:
\describe{
\item{\code{"default"}}{The default value and will render and return the
following:
\enumerate{
\item SOW: All SOW language in one ".docx" file.
\item CDRLs: All cost reporting CDRLs in one ".docx" file.
}}
\item{\code{"cade_ready"}}{Returns the same files as \code{"default"}
but splits the CDRL document into multiple files (each CSDR CDRL gets
its own file and each non-CSDR CDRL gets lumped into one file) for
easier uploading into the CADE's Program Planning Module.
}}}
}
\value{
Returns nothing to R but does puts the following files at the
\code{output_dir}, dependent on what is passed to the
\code{files_to_render} paramater:
\enumerate{
\item A Word document with cost reporting scope of work language,
filled in with the param values from the "cost_reporting_params.xlsx"
file.
\item A Word document or documents with cost reporting CDRLs, filled
in with the param values from the "cost_reporting_params.xlsx" file.
\item A partially complete CSDR Plan attachment, filled in with the
param  values from the "cost_reporting_params.xlsx" file.
\item A partially complete RDT attachment, filled in with the param
values from the "cost_reporting_params.xlsx" file.
\item A partially complete LSPD Report attachment (if required),
filled in with the param values from the "cost_reporting_params.xlsx"
file.
\item A complete AUMC Report attachment (if required), filled in with
the param values from the "cost_reporting_params.xlsx" file.
}
}
\description{
Imports a pre-formatted and filled in Excel file with info on the planned
contract and system then generates a scope of work, CDRLs, and associated
contract attachments from template documents.
}
\examples{
\dontrun{
# Render the cost reporting documents and select both the
# "cost_reporting_params.xlsx" file and output directory using pop-up boxes.
render_cr_docs()

# Render the cost reporting documents with the default
# "cost_reporting_params.xlsx" values and put the output on a fake user's
# desktop.
render_cr_docs(params_path = here::here("extdata", "cost_reporting_params.xlsx"),
               output_dir = "C:/Users/john.doe/Desktop")
         }
}
