% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_cr_docs.R
\name{render_cr_docs}
\alias{render_cr_docs}
\title{Create Cost Reporting Scope of Work from Template}
\usage{
render_cr_docs(params_path = file.choose(), output_dir = choose.dir())
}
\arguments{
\item{params_path}{Required. A full path to a filled-in
"cost_reporting_params.xlsx" file. If a pass is not passed to the function
call, a file selection box will be opened.}

\item{output_dir}{Required. A directory path to where the user wants the
generated files to be placed. If a path is not passed by Windows users, a
pop-up box will appear. Non-Windows users must pass a directory path.}
}
\value{
Returns nothing to R but does put the following files at t \code{output_dir}:
\enumerate{
\item A Word document with cost reporting scope of work language.
\item A Word document with cost reporting CDRL language.
\item A partially complete CSDR Plan attachment.
\item A partially complete RDT attachment.
\item A partially complete LSPD Report attachment (if required).
\item A complete AUMC Report attachment (if required).
}
}
\description{
Imports a pre-formatted and filled in Excel file with info on the planned
contract and system then generates a scope of work, CDRLs, and associated
contract attachments from template documents.
}
\examples{
\dontrun{
render_cr_docs(
  params_path = here::here("inst", "cost_reporting_params.xlsx"),
  output_dir = here::here()
)
}
}
