% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_cr_docs.R
\name{render_cr_docs}
\alias{render_cr_docs}
\title{Create Cost Reporting SOW, CDRLs, Attachments with Params}
\usage{
render_cr_docs(params_path = file.choose(), output_dir = choose.dir())
}
\arguments{
\item{params_path}{Required. A full path to a filled-in
"cost_reporting_params.xlsx" file. If a path is not provided to the
function call, a file selection box will be opened.}

\item{output_dir}{Required. A directory path to where the user wants the
generated files to be placed. If a path is not passed by Windows users, a
pop-up box will appear. Non-Windows users must pass a directory path.}
}
\value{
Returns nothing to R but does put the following files at the
\code{output_dir}: \enumerate{ \item A Word document with cost reporting
scope of work language, filled in with the param values from the
"cost_reporting_params.xlsx" file. \item A Word document with cost
reporting CDRLs, filled in with the param values from the
"cost_reporting_params.xlsx" file. \item A partially complete CSDR Plan
attachment, filled in with the param values from the
"cost_reporting_params.xlsx" file. \item A partially complete RDT
attachment, filled in with the param values from the
"cost_reporting_params.xlsx" file. \item A partially complete LSPD Report
attachment (if required), filled in with the param values from the
"cost_reporting_params.xlsx" file. \item A complete AUMC Report attachment
(if required), filled in with the param values from the
"cost_reporting_params.xlsx" file. }
}
\description{
Imports a pre-formatted and filled in Excel file with info on the planned
contract and system then generates a scope of work, CDRLs, and associated
contract attachments from template documents.
}
\examples{
\dontrun{
render_cr_docs(
  params_path = here::here("extdata", "cost_reporting_params.xlsx"),
  output_dir = here::here()
)}
}
